+++ instrument-panel/instruments/annunciator.cpp
@@ -230,7 +230,7 @@ void annunciator::update()
     }
 
     // Check for position or size change
-    long *settings = globals.simVars->readSettings(name, xPos, yPos, size);
+    int *settings = globals.simVars->readSettings(name, xPos, yPos, size);
 
     xPos = settings[0];
     yPos = settings[1];
